requirements.txt
alembic==1.11.3
Babel==2.13.0
blinker==1.6.2
click==8.1.7
colorama==0.4.6
css-html-js-minify==2.5.5
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
flask-babel==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
WTForm==1.0
WTForms==3.0.1
